+++ CHANGELOG.md
@@ -5,3 +5,4 @@ v0.2
 ----
 - Gracefully handle ^C
 - Package using poetry
+- Add -n option for limiting search results